Rename along vue style guide
