+++ template/components/AppLogo.vue
